Workaround for #190 and #207
For some reason, when Package Control updates Anaconda in Windows, the
winsocks enter in an inconsistent state that produce several problems,
being the most critical one a total unresponsiveness of the ST3.

What Package Control do is just disable the package, install the update
and then re-enable de package, if you do that manually, it seems to work
just fine. I am being able to reproduce this behaviour after anaconda
automatic updates only.

The commit just detect when a socket enter in inconsistent state, then
show an error message to the user and stop the asynchronous poll (this
should not make ST3 unresponsive).