Bugfix: Additional py.test support for Test Runner
py.test requires the following format when specifying targetted tests:

		py.test path/to/file.py::ModuleName::test_name

Because of this, the `Run Current Test` and `Run Current File Tests` commands
produced test paths that were incompatible with py.test.

This commit adds:

* a `test_filepath_patterns` setting that tells the Test Runner
to use filepaths when building test patterns paths.

* an additional delimeter setting named `test_method_delimeter` that allows
the user to specify the delimeter to be used to separate the targeted class
and method.

Recommended settings for py.test users:
* test_delimeter: "::"
* test_method_delimeter: "::"
* test_filepath_patterns: true